Created Overrides module for handling externally defined settings.
