+++ lib/plugins/branches.js
@@ -1,10 +1,11 @@
 const ErrorStash = require('./errorStash')
 const NopCommand = require('../nopcommand')
 const MergeDeep = require('../mergeDeep')
+const Overrides = require('./overrides')
 const ignorableFields = []
 const previewHeaders = { accept: 'application/vnd.github.hellcat-preview+json,application/vnd.github.luke-cage-preview+json,application/vnd.github.zzzax-preview+json' }
 const overrides = [
-  'required_status_checks.contexts',
+  'contexts',
 ]
 
 module.exports = class Branches extends ErrorStash {
@@ -52,7 +53,7 @@ module.exports = class Branches extends ErrorStash {
               const params = Object.assign({}, p)
               return this.github.repos.getBranchProtection(params).then((result) => {
                 const mergeDeep = new MergeDeep(this.log, this.github, ignorableFields)
-                const changes = mergeDeep.compareDeep({ branch: { protection: this.reformatAndReturnBranchProtection(result.data) } }, { branch: { protection: this.removeOverrides(branch.protection, result.data) } })
+                const changes = mergeDeep.compareDeep({ branch: { protection: this.reformatAndReturnBranchProtection(result.data) } }, { branch: { protection: Overrides.removeOverrides(overrides, branch.protection, result.data) } })
                 const results = { msg: `Followings changes will be applied to the branch protection for ${params.branch.name} branch`, additions: changes.additions, modifications: changes.modifications, deletions: changes.deletions }
                 this.log.debug(`Result of compareDeep = ${results}`)
 
@@ -79,7 +80,7 @@ module.exports = class Branches extends ErrorStash {
                 return this.github.repos.updateBranchProtection(params).then(res => this.log(`Branch protection applied successfully ${JSON.stringify(res.url)}`)).catch(e => { this.logError(`Error applying branch protection ${JSON.stringify(e)}`); return [] })
               }).catch((e) => {
                 if (e.status === 404) {
-                  Object.assign(params, this.removeOverrides(branch.protection, {}), { headers: previewHeaders })
+                  Object.assign(params, Overrides.removeOverrides(overrides, branch.protection, {}), { headers: previewHeaders })
                   if (this.nop) {
                     resArray.push(new NopCommand(this.constructor.name, this.repo, this.github.repos.updateBranchProtection.endpoint(params), 'Add Branch Protection'))
                     return Promise.resolve(resArray)
@@ -124,34 +125,5 @@ module.exports = class Branches extends ErrorStash {
       }
     }
     return protection
-  }
-
-  getLastKey (obj, property) {
-    const keys = property.split('.')
-    for (let i = 0; i < keys.length - 1; i++) {
-      if (!obj[keys[i]]) return [null, null]
-      obj = obj[keys[i]]
-    }
-    return [obj, keys[keys.length - 1]]
-  }
-
-  removeOverrides (source, existing) {
-    if (source) {
-      overrides.forEach(override => {
-        let [obj, lastKey] = this.getLastKey(source, override)
-        if (!obj) return
-        if ((Array.isArray(obj[lastKey]) || typeof obj[lastKey] === 'string') && obj[lastKey].includes('{{EXTERNALLY_DEFINED}}')) {
-          let [obj2, lastKey2] = this.getLastKey(existing, override)
-          if (obj2[lastKey2]) {
-            obj[lastKey] = obj2[lastKey2]
-          } else if (Array.isArray(obj[lastKey])) {
-            obj[lastKey] = []
-          } else {
-            obj[lastKey] = ''
-          }
-        }
-      })
-    }
-    return source
   }
 }
