+++ lib/plugins/rulesets.js
@@ -1,6 +1,7 @@
 const Diffable = require('./diffable')
 const NopCommand = require('../nopcommand')
 const MergeDeep = require('../mergeDeep')
+const Overrides = require('./overrides')
 const ignorableFields = []
 const overrides = [
   'required_status_checks',
@@ -95,51 +96,6 @@ module.exports = class Rulesets extends Diffable {
     return merged.hasChanges
   }
 
-  getObjectRef (data, dataKey) {
-    const results = []
-    const traverse = (obj) => {
-      for (const key in obj) {
-        if (key === dataKey) {
-          results.push(obj)
-        } else if (Array.isArray(obj[key])) {
-          obj[key].forEach(element => traverse(element))
-        } else if (typeof obj[key] === 'object' && obj[key]) {
-          traverse(obj[key])
-        }
-      }
-    }
-    traverse(data)
-    return results
-  }
-
-  // When {{EXTERNALLY_DEFINED}} is found in the override value, retain the
-  // existing value from GitHub.
-  // Note:
-  //   - The admin settings could define multiple overrides, but the GitHub API
-  //     retains one only.
-  //   - The PUT method for rulesets (update) allows for multiple overrides.
-  //   - The POST method for rulesets (create) allows for one override only.
-  removeOverrides (source, existing) {
-    overrides.forEach(override => {
-      let sourceRefs = this.getObjectRef(source, override)
-      let data = JSON.stringify(sourceRefs, null, 2)
-      if (data.includes('{{EXTERNALLY_DEFINED}}')) {
-        let existingRefs = this.getObjectRef(existing, override)
-        sourceRefs.forEach(sourceRef => {
-          if (existingRefs[0]) {
-            sourceRef[override] = existingRefs[0][override]
-          } else if (Array.isArray(sourceRef[override])) {
-            sourceRef[override] = []
-          } else if (typeof sourceRef[override] === 'object' && sourceRef[override]) {
-            sourceRef[override] = {}
-          } else {
-            sourceRef[override] = ''
-          }
-        })
-      }
-    })
-  }
-
   update (existing, attrs) {
     const parms = this.wrapAttrs(Object.assign({ id: existing.id }, attrs))
     if (this.scope === 'org') {
@@ -148,7 +104,7 @@ module.exports = class Rulesets extends Diffable {
           new NopCommand(this.constructor.name, this.repo, this.github.request.endpoint('PUT /orgs/{org}/rulesets/{id}', parms), 'Update Ruleset')
         ])
       }
-      this.removeOverrides(parms, existing)
+      Overrides.removeOverrides(overrides, parms, existing)
       this.log.debug(`Updating Ruleset with the following values ${JSON.stringify(parms, null, 2)}`)
       return this.github.request('PUT /orgs/{org}/rulesets/{id}', parms).then(res => {
         this.log(`Ruleset updated successfully ${JSON.stringify(res.url)}`)
@@ -162,7 +118,7 @@ module.exports = class Rulesets extends Diffable {
           new NopCommand(this.constructor.name, this.repo, this.github.request.endpoint('PUT /repos/{owner}/{repo}/rulesets/{id}', parms), 'Update Ruleset')
         ])
       }
-      this.removeOverrides(parms, existing)
+      Overrides.removeOverrides(overrides, parms, existing)
       this.log.debug(`Updating Ruleset with the following values ${JSON.stringify(parms, null, 2)}`)
       return this.github.request('PUT /repos/{owner}/{repo}/rulesets/{id}', parms).then(res => {
         this.log(`Ruleset updated successfully ${JSON.stringify(res.url)}`)
@@ -180,7 +136,7 @@ module.exports = class Rulesets extends Diffable {
           new NopCommand(this.constructor.name, this.repo, this.github.request.endpoint('POST /orgs/{org}/rulesets', this.wrapAttrs(attrs)), 'Create Ruleset')
         ])
       }
-      this.removeOverrides(attrs, {})
+      Overrides.removeOverrides(overrides, attrs, {})
       this.log.debug(`Creating Rulesets with the following values ${JSON.stringify(attrs, null, 2)}`)
       return this.github.request('POST /orgs/{org}/rulesets', this.wrapAttrs(attrs)).then(res => {
         this.log(`Ruleset created successfully ${JSON.stringify(res.url)}`)
@@ -194,7 +150,7 @@ module.exports = class Rulesets extends Diffable {
           new NopCommand(this.constructor.name, this.repo, this.github.request.endpoint('POST /repos/{owner}/{repo}/rulesets', this.wrapAttrs(attrs)), 'Create Ruleset')
         ])
       }
-      this.removeOverrides(attrs, {})
+      Overrides.removeOverrides(overrides, attrs, {})
       this.log.debug(`Creating Rulesets with the following values ${JSON.stringify(attrs, null, 2)}`)
       return this.github.request('POST /repos/{owner}/{repo}/rulesets', this.wrapAttrs(attrs)).then(res => {
         this.log(`Ruleset created successfully ${JSON.stringify(res.url)}`)
